aspen-mcp/requirements.txt
mcp[cli]>=1.0.0
pywin32>=306
pyyaml>=6.0
